delete old files and rename demo
