+++ demo/multivalue.html
@@ -1,0 +1,47 @@
+<!DOCTYPE html>
+<html>
+<head lang="en">
+  <meta charset="UTF-8">
+  <title>LineUp alpha</title>
+
+  <link href="../node_modules/font-awesome/css/font-awesome.css" rel="stylesheet" type='text/css'>
+
+  <link href="../build/style.css" rel="stylesheet">
+
+  <style>
+    div.orange {
+      background-color: orange;
+    }
+
+    rect.bar.orange {
+      fill: orange;
+    }
+
+    div.green {
+      background-color: green;
+    }
+
+    rect.bar.green, .bar.green rect {
+      fill: green;
+    }
+
+    .lu {
+      clear: both;
+      position: absolute;
+      top: 1px;
+      bottom: 0;
+      left: 0;
+      right: 0;
+      padding: 0;
+      overflow: auto;
+    }
+
+  </style>
+</head>
+<body>
+<script src="../node_modules/d3/d3.js"></script>
+<script src="../build/LineUpJS.js"></script>
+<script src="./multivalue.js"></script>
+
+</body>
+</html>
